Make thin_number a u32
dm-thin only allows 24 bit thin number.

+++ src/main.rs
@@ -512,7 +512,7 @@ impl ThinPoolDev {
 #[derive(Debug)]
 struct ThinDev {
     dev: Device,
-    thin_number: u64,
+    thin_number: u32,
 }
 
 impl ThinDev {
